requirements.txt
absl-py==0.12.0
aiofiles==0.5.0
aiohttp==3.7.4.post0
aniso8601==7.0.0
APScheduler==3.6.3
astunparse==1.6.3
async-exit-stack==1.0.1
async-generator==1.10
async-lru==1.0.2
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.32.0
h11==0.12.0
h5py==2.10.0
httptools==0.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
motor==2.4.0
multidict==5.1.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
orjson==3.5.2
pandas==1.2.4
Pillow==8.2.0
prettytable==2.1.0
promise==2.3
protobuf==3.15.8
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.1
pymongo==3.11.4
pyparsing==2.4.7
Pyrogram==1.2.9
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.0
python-multipart==0.0.5
python-telegram-bot==13.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==1.6.1
scikit-learn==0.22.2.post1
scipy==1.6.3
seaborn==0.11.1
six==1.15.0
starlette==0.13.6
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
TgCrypto==1.2.2
threadpoolctl==2.1.0
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==3.2.0
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
